Extracted Markdown-It plugin, made plugin testable
