+++ src/test/markdown-it/index.ts
@@ -1,0 +1,39 @@
+import * as path from "path";
+import Mocha from "mocha";
+import { glob } from "glob";
+
+export async function run(): Promise<void> {
+  // Create the mocha test
+  const mocha = new Mocha({
+    ui: "tdd",
+    color: true,
+  });
+
+  const testsRoot = path.resolve(__dirname, ".");
+  console.log("Tests Root", testsRoot);
+  const files = await glob("**/**.test.js", { cwd: testsRoot });
+  console.log("Test files", files);
+
+  // Add files to the test suite
+  files.forEach((f) => mocha.addFile(path.resolve(testsRoot, f)));
+
+  try {
+    return new Promise<void>((c, e) => {
+      // Run the mocha test
+      mocha.run((failures) => {
+        console.log("Testing complete");
+        if (failures > 0) {
+          e(new Error(`${failures} tests failed.`));
+        } else {
+          c();
+        }
+      });
+    });
+  } catch (err) {
+    console.error(err);
+  }
+}
+
+
+// RUN TESTS
+run();
