+++ src/test/markdown-it/scriptPlugin.test.ts
@@ -1,0 +1,9 @@
+import * as assert from "assert";
+
+// TODO: use jest
+suite("Markdown-It plugin tests", () => {
+  test("test rendering", () => {
+    assert.strictEqual(-1, [1, 2, 3].indexOf(5));
+    assert.strictEqual(1, [1, 2, 3].indexOf(0));
+  });
+});
