+++ src/test/suite/extension.test.ts
@@ -10,6 +10,6 @@ suite('Extension Test Suite', () => {
 
 	test('Sample test', () => {
 		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
+		assert.strictEqual(0, [1, 2, 3].indexOf(0));
 	});
 });
